analysis/python/requirements.txt
pandas>=2.0
numpy>=1.24
lightgbm>=4.0
scikit-learn>=1.3
matplotlib>=3.7
requests>=2.31
pyyaml>=6.0
joblib>=1.3
pytest>=7.0
